Fixed warnings on Solaris with Sun's C compiler.
git-svn-id: file:///home/survex-svn/survex/trunk@806 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/netbits.c
@@ -245,7 +245,7 @@ addfakeleg(node *fr, node *to,
 #endif
 	   )
 {
-   uchar i, j;
+   int i, j;
    linkfor *leg, *leg2;
    int shape;
